requirements.txt
# Skill Router Dependencies
httpx>=0.27.0
numpy>=1.24.0
python-dotenv>=1.0.0
